Unified component tables + removed palette system + fixed UTF-8 BOM
MAJOR CHANGES:
- Unified component tables: Single p32_component_tables.h replaces 3 separate headers
- Single COMPONENT_TABLE_SIZE constant (replaces INIT/ACT/HITCOUNT_TABLE_SIZE)
- Removed obsolete palette-based rendering (p32_mood_palette.*, p32_mood_transitions.*)
- Using direct RGB565 pixel manipulation (FrameProcessor) for performance
- Fixed UTF-8 BOM corruption in goblin_full.json that broke JSON parser
- Added RULE 12 to AI-AGENT-RULES.md: JSON files must be ASCII (no BOM)
- Regenerated component tables from goblin_full.json (9 components)
- Updated CMakeLists.txt to include system_components.c and bot_components.c
- Added g_loopCount global to component tables

COMPONENTS (9 total):
- System: heartbeat, network_monitor
- Bot: left_eye, right_eye, mouth, speaker, nose_sensor, left_ear_microphone, right_ear_microphone

BUILD STATUS:  Compiles successfully

FILES ADDED:
- .github/AI-AGENT-RULES.md (ironclad rules for AI agent)
- build.ps1 (clean + build script)
- include/p32_component_tables.h (unified header)
- Multiple hardware validation docs

FILES DELETED:
- include/actTable.h, initTable.h (replaced by unified header)
- include/p32_mood_palette.h (obsolete palette system)
- src/actTable.c, initTable.c (replaced by unified implementation)
- src/p32_mood_palette.c, p32_mood_transitions.c (obsolete)

CRITICAL BUG FIX:
- UTF-8 BOM (bytes EF BB BF) in JSON files breaks Python json.load()
- Error: 'Expecting value: line 1 column 1 (char 0)'
- Solution: Save all JSON as pure ASCII encoding
- RULE 12 enforces this to prevent recurrence

+++ src/p32_shared_state.cpp
@@ -4,8 +4,10 @@
 #include "p32_shared_state.h"
 #include <string.h>
 
-// Global loop counter (increments every 100ms in main loop)
-uint32_t g_loopCount = 0;
+// Global loop counter (64-bit to prevent overflow at 120,000 iterations/sec)
+// NOTE: Non-const definition here, allows main.c to write via extern declaration
+// Components see it as const via p32_shared_state.h
+uint64_t g_loopCount = 0;
 
 // Global shared state (synchronized via ESP-NOW mesh across all subsystems)
 p32_shared_state_t g_shared_state = {
